requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jinja2
flask
requests
python-dotenv
google-genai